Fix wrong date formatting (#164)
Fix wrong date formatting in some cases. For example when the specified date format in WP is 'd/m/Y' then strtotime will parse a date like '05/11/2017' as 'm/d/Y' (see strtotime for details) which causes the month and date to be flipped. Instead, we should format the date in a specific accepted format (for example 'Y-m-d H:i:s') to prevent this problem. Afterwards it is correctly formatted for display using date_i18n and wc_date_format anyway.

Sidenote: I have not tested the change, but it should work.