WIP: Automated deployment of the plugin to Wordpress
This branch will contain an updated deployment script that automatically prepares a new version of the WooCommerce plugin and releases it to the Wordpress store. The base deployment script that has been adjusted can be seen in this repository: https://github.com/GaryJones/wordpress-plugin-svn-deploy

Modifications made to fit the script to our plugin structure:
- Replaced input prompts by static values.
- Remove all the files from the Wordpress SVN trunk before checkout out all the plugin files, otherwise the checkout proces will not execute successfully.
- Install Composer in the script and automatically install all the dependencies of the libraries that the WooCommerce plugin requires, by searching for all `composer.json` files.
- Automatically ignore all files we don't want to commit to SVN, which is defined in `deployment/.svnignore`.

The script is still WIP, and it is halted before actually committing the changes to SVN. All files seem to be in place correctly, except for that one `mollie-api-php` endpoint class and example are deleted for some reason. To see the changes that would be committed to SVN, please run the following commands from the root of this repository, after running the deployment script:
```
cd deployment/mollie-payments-for-woocommerce
svn status
```

After verifying that the file structure is as it should be, we should probably test locally that the committing process actually works, and then configure Travis to run this when a new tag is pushed to master, using the correct credentials.