Introduce Apple Pay payment method
Because WooCommerce render the payment methods twice, one when the
page is rendered by php another one by doing an ajax call to one of
its endpoints I added an hidden field in the checkout page to check against
the iDevice and set the hidden input in case is not possible to show
the apple pay payment, this will be checked by `addGateways` method
that will remove the Apple Pay gateway.

So basically the solution will hide the payment method from the frontend
where the backend still continue to have the gateway available.