Remove Apple Pay method securely without causing type errors
WooCommerce allow to register gateways not only as class string names but
you can also register an instance of a class.

When we are looking for `Mollie_WC_Gateway_Applepay` to remove it using
`array_diff` we could cause an error because the `$gateways` list may
contains instances of classes.

Using array search will fix the problem.

The solution is unit tested