Always check if active payment instance has been set properly
The method `getActiveMolliePayment` could return two different type of
values, this make the solution unpredictable and you have to add additional
conditional statement.

Instead a more correct approach would be throw an ´UnexpectedValueException`.

The main problem was to set a different message for caller about the
return value of `getActiveMolliePayment` that could help debugging and
support and also solve a WSOD in production.

Problem was the impossibility to create an unit tested solution.
To do that I created a new method for `paymentObject` and `activePaymentObject`
which move the concern of create instances outside of a consumer method.

Since I while working on that I found some duplication code that could
be improved and I also changed the way the order id is retrieved in a more
DRY version.

Now instead of duplicate if/else we have a method that will encapsulate the
BC logic.

Some small improvements are made to the use of string interpolation over
concatenation and removing duplicated WC_Order assignments.