Centralized base path constants and added main IIFE
Plugin dir is determined from the plugin main file path, which
is determined _once_ from the location of the main file. This
creates a single source of truth for base paths. This source
of truth is now used throughout the main file, including the
new closure, which loads Composer's autoload file, instead
of custom stuff.