Refund chargeback the merchant for every order line item
The issue also solve a problem when the amount to refund for an item
is different than the price per unit defined in the mollie order.

# Deprecate action
Deprecate `Mollie_WC_Plugin::PLUGIN_ID . '_refund_created'` action, in
place of that there are multiple different actions performed based on
a specific context.

# Wrap getPaymentObject in a try/catch block
The changes introduce some DI expecially for `orders` endpoint.
Since `orders` endpoing are retrieved from the `ApiClient` when the
`paymentObject` is constructed it's possible that some of the internals
will throw an `ApiException`, for this reason I wrap the construction
of the object in a try/catch block.

Every catch block will log the exception but the returned value depends
by the context.

# Improve `Mollie_WC_Gateway_AbstractSepaRecurring::handlePaidOrderWebhook`
Prevent useless and complicated to maintain code duplication just for
BC compatibility for WooCommerce.

# Avoid useless try/catch block
There's a try/catch block which just throw again the same exception
without doing nothing useful in `Helper::getApiClient`.

MOL-33