Remove code coverage since we do not have any place where we can manage it
We need a place where it is possible to see the code coverage without
having to run all of the tests everytime.

Right now we do not have that place and also the coverage of the code is
ridiculous and also seems in php7.4 in travis they are missing something
such as xdebug because tests wont run because of the following error

`Error:         No code coverage driver is available`

So since actually code coverage is not necessary I'm removing the parameter
from travic phpunit command.