Temporary disable notice at checkout level if token cannot be retrieved
Sounds like mollie components have problems in unmount components.
The main problem is caused because WooCommerce has a js event `updated_checkout`
which is fired when an input field change and the markup of the payment methods
is recreated from scratch every time.

We have to unmount the components or the following error will thrown.

```
mollie.js:1 Uncaught MollieComponentsError: Cannot create cardHolder component because a component of the same type already exists.
```

But seems that even unmount does not help us in mount new components again.
So, since mollie will show the components fields on their payment page
for now we disable the notice, at least in case of problems the customer
can still pay.