ApplePay button feature
MOL-95 MOL-146

Adds a new ApplePay button on product detail page
and cart page if the feature is enabled.
Checks for https and bootstraps the feature by
printing the button.
On clicking, the workflow is a round of ajax
calls:
validate merchant: against Apple through Mollie,
update shipping contact: or correct errors,
update shipping method:or correct errors. Both
also update the amount to pay if needed.
On payment authorized by the user, the order is
created and sent to Mollie with the given token,
so is registered in WooCommerce as usual.
The required validation file is served as a string
when the set url is called so user does not have
to deal with new files on the server.
Notices appear on failed payment and on validation
error. Logs register missing data and errors.
Added tests for success and errors.