Add bank transfer expiration date setting
MOL-182
Two new settings, one enables the feature, the other allows the user
to choose the due date in days. If the feature is enabled all payments
will be treated as payment not orders, so a method to check this was
added to process_payment method. Added small refactors for all the
involved methods.