Include functions in all tests
This ensures that the functions are available in all of the tests automatically, and that mocks are not created before the functions have been included. Without this, running multiple tests that mock functions in the same process results in an error, because it tries to include functions which already exist due to being mocked.