Add missing import
Without it, a new class was created for the mock, without inheriting from the real class. This caused some software checks to fail.
To make it work, I needed to remove the explicit mock class name (because that class will have already existed, since the actual class has now been imported. I also had to create a mock for abstract class, because it is abstract.