Run specific test in separate process
This is necessary because the things it is testing rely on values that usually change between requests. Having this test in the same process as other tests would make the files that should not be enqueued during the request simulated by the tests enqueued anyway - perhaps due to logic in a different test.