Will now download extra files for PHPScoper
PHPScoper does not have support for WP out of the box. This is because WP is not actually present in the plugin, and therefore PHPScoper cannot understand that the WP functions are global, and to prefix and alias them properly.

The WP stubs - empty WP symbol declarations - are downloaded, as well as symbol extractor and patcher. In conjunction, these can be used by PHPScoper to un-prefix WordPress symbols. It's hacky, but probably a temporary measure. A solution is being discussed:
https://github.com/humbug/php-scoper/issues/303