Configure PHPScoper to leave WP symbols alone
This uses the WP stubs, the identifier extractor, and the prefix remover together. The latter two cannot be required with composer, because during build only production dependencies are installed and these are not needed in production. Because of this, their class files must be included manually in PHP.

Also, a class alias is added for `PhpParser/ParserFactory`, because PHPScoper itself is scoped inside the Phar, and this associates the scoped class name with the original. It's necessary because the `IdentifierExtractor` references the original classname.