Consistent transaction descriptions
Since it is now always attempted to create an order it would be nice to have consistent transaction.description. For transactions via the orders API this will be the "Order "+ order.orderNumber. However when this is done via the payments API it is a translation of "Order" done via the WP method __() + orderNumber. I removed this translation.