Add api selector setting
Extract form fields to settings file
Add new setting to switch API order/payments
Default is order, then we check the setting,
then we check the gateway is banktransfer
(additional setting) then we check the products.
If the API is payments the merchant can choose a
description, so we use it in the request.