MOL-891 remove some coupling with wc subs
Some methods check if there is a subscription plugin.
Now that check can be true for plugins declaring themselves as subscription via filter.
DirectDebit gateway is enabled upon that same filter.
Not only if WC_Subscriptions class is present.
The methods where we handle WC_Subs code still need to check for it being present.